rollback move of some free functions sound more intuitive to keep them into install instead of create header
