+++ src/api/update.cpp
@@ -5,7 +5,7 @@
 // The full license is in the file LICENSE, distributed with this software.
 
 #include "mamba/api/configuration.hpp"
-#include "mamba/api/create.hpp"
+#include "mamba/api/install.hpp"
 #include "mamba/api/update.hpp"
 
 #include "mamba/core/context.hpp"
